+++ mshell/Lexer.go
@@ -31,6 +31,7 @@ const (
 	MINUS
 	PLUS
 	EQUALS
+	FATARROW
 	INTERPRET
 	IF
 	IFF // This is just temporary as I work to remove the old if.
@@ -98,6 +99,7 @@ const (
 	GRID_CLOSE               // |]
 	PREFIXQUOTE              // .functionName
 	MATCH
+	UNPACK
 	VER
 	STDERRTOSTDOUT           // 2>&1, merge stderr into stdout's destination
 	STDOUTTOSTDERR           // 1>&2, merge stdout into stderr's destination
@@ -149,6 +151,8 @@ func (t TokenType) String() string {
 		return "PLUS"
 	case EQUALS:
 		return "EQUALS"
+	case FATARROW:
+		return "FATARROW"
 	case INTERPRET:
 		return "INTERPRET"
 	case IF:
@@ -283,6 +287,8 @@ func (t TokenType) String() string {
 		return "PREFIXQUOTE"
 	case MATCH:
 		return "MATCH"
+	case UNPACK:
+		return "UNPACK"
 	case VER:
 		return "VER"
 	case AS:
@@ -641,6 +647,8 @@ func (l *Lexer) literalOrKeywordType() TokenType {
 				return l.checkKeyword(2, "pe", TYPE)
 			}
 		}
+	case 'u':
+		return l.checkKeyword(1, "npack", UNPACK)
 	case 'V':
 		return l.checkKeyword(1, "ER", VER)
 	case 'x':
@@ -784,6 +792,10 @@ func (l *Lexer) scanTokenAll() Token {
 			return l.parseLiteralOrKeyword()
 		}
 	case '=':
+		if l.peek() == '>' {
+			l.advance()
+			return l.makeToken(FATARROW)
+		}
 		return l.makeToken(EQUALS)
 	case ',':
 		return l.makeToken(COMMA)